006demo/requirements.txt
langchain
langgraph
langchain-openai
mcp
python-dotenv
openai
httpx
beautifulsoup4
